refactor(system-transformations): isolate, implememnt interface in VizPanel
